refactor: 통합 테스트 코드 리팩토링
테스트 환경 구성을 추상 클래스로 분리하여 재사용 할 수 있도록 함.
